feat: add dropzone to preview
