+++ components/Forms/DropzoneInput.tsx
@@ -1,18 +1,9 @@
-import { MouseEvent, ReactElement, useCallback, useState } from 'react';
+import clsx from 'clsx';
+import { MouseEvent, useCallback } from 'react';
 import { useFormContext } from 'react-hook-form';
 import { FileWithPath, useDropzone } from 'react-dropzone';
-import Lightbox from 'react-image-lightbox';
-import clsx from 'clsx';
 
-import {
-  HiOutlineExternalLink,
-  HiOutlineEye,
-  HiOutlinePaperClip,
-  HiOutlinePhotograph,
-  HiX,
-} from 'react-icons/hi';
-
-import UnstyledLink from '@/components/UnstyledLink';
+import { FilePreview } from './FilePreview';
 
 type DropzoneInputProps = {
   accept?: string;
@@ -24,101 +15,6 @@ type DropzoneInputProps = {
   validation?: object;
 };
 
-type FilePreviewProps = {
-  deleteFile: (e: MouseEvent, file: FileWithPath) => void;
-  file: FileWithPath;
-};
-
-const FilePreview = ({ deleteFile, file }: FilePreviewProps): ReactElement => {
-  const [index, setIndex] = useState(0);
-  const [isOpen, setIsOpen] = useState(false);
-
-  const images = [URL.createObjectURL(file)];
-
-  const handleDelete = (e: MouseEvent) => {
-    e.stopPropagation();
-    deleteFile(e, file);
-  };
-
-  const imagesType = ['image/png', 'image/jpg', 'image/jpeg'];
-
-  return imagesType.includes(file.type) ? (
-    <>
-      <li
-        className='flex items-center justify-between py-3 pl-3 pr-4 text-sm'
-        key={file.name}
-      >
-        <div className='flex items-center flex-1 w-0'>
-          <HiOutlinePhotograph
-            className='flex-shrink-0 w-5 h-5 text-gray-400'
-            aria-hidden='true'
-          />
-          <span className='flex-1 w-0 ml-2 truncate'>{file.name}</span>
-        </div>
-        <div className='flex-shrink-0 ml-4'>
-          <button
-            type='button'
-            onClick={() => setIsOpen(true)}
-            className='inline-block mr-2 text-xl font-medium text-gray-500 focus:outline-none hover:text-gray-700'
-          >
-            <HiOutlineEye />
-          </button>
-          <button
-            type='button'
-            onClick={handleDelete}
-            className='text-xl font-medium text-red-500 focus:outline-none hover:text-red-700'
-          >
-            <HiX />
-          </button>
-        </div>
-      </li>
-      {isOpen && (
-        <Lightbox
-          mainSrc={images[index]}
-          nextSrc={images[(index + 1) % images.length]}
-          prevSrc={images[(index + images.length - 1) % images.length]}
-          onCloseRequest={() => setIsOpen(false)}
-          onMovePrevRequest={() =>
-            setIndex(
-              (prevIndex) => (prevIndex + images.length - 1) % images.length
-            )
-          }
-          onMoveNextRequest={() =>
-            setIndex((prevIndex) => (index + 1) % images.length)
-          }
-        />
-      )}
-    </>
-  ) : (
-    <li
-      key={file.name}
-      className='flex items-center justify-between py-3 pl-3 pr-4 text-sm'
-    >
-      <div className='flex items-center flex-1 w-0'>
-        <HiOutlinePaperClip
-          className='flex-shrink-0 w-5 h-5 text-gray-400'
-          aria-hidden='true'
-        />
-        <span className='flex-1 w-0 ml-2 truncate'>{file.name}</span>
-      </div>
-      <div className='flex space-x-2'>
-        <UnstyledLink href={URL.createObjectURL(file)}>
-          <HiOutlineExternalLink
-            size={24}
-            className='text-gray-500 hover:text-gray-700'
-          />
-        </UnstyledLink>
-        <button type='button' onClick={(e) => deleteFile(e, file)}>
-          <HiX
-            size={24}
-            className='text-red-500 cursor-pointer hover:text-red-700'
-          />
-        </button>
-      </div>
-    </li>
-  );
-};
-
 export default function DropzoneInput({
   accept,
   helperText = '',
@@ -126,6 +22,7 @@ export default function DropzoneInput({
   label,
   maxFiles = 1,
   validation,
+  readOnly,
 }: DropzoneInputProps) {
   const {
     register,
@@ -150,7 +47,7 @@ export default function DropzoneInput({
         clearErrors(id);
       }
     },
-    [id, setValue, setError, clearErrors]
+    [setValue, id, setError, clearErrors]
   );
 
   const deleteFile = (e: MouseEvent, file: FileWithPath) => {
@@ -179,22 +76,34 @@ export default function DropzoneInput({
         {label}
       </label>
 
-      {files?.length >= maxFiles ? (
-        <ul className='border border-gray-200 divide-y divide-gray-200 rounded-md'>
+      {readOnly && !(files?.length > 0) ? (
+        <div className='py-3 pl-3 pr-4 text-sm border border-gray-300 divide-y divide-gray-300 rounded-md'>
+          No file uploaded
+        </div>
+      ) : files?.length >= maxFiles ? (
+        <ul className='border border-gray-300 divide-y divide-gray-300 rounded-md'>
           {files.map((file, index) => (
-            <FilePreview key={index} file={file} deleteFile={deleteFile} />
+            <FilePreview
+              key={index}
+              readOnly={readOnly}
+              file={file}
+              deleteFile={deleteFile}
+            />
           ))}
         </ul>
       ) : (
         <>
-          <div className='mt-1' {...getRootProps()}>
+          <div
+            className='mt-1 focus:outline-none focus:ring-dark-400 group'
+            {...getRootProps()}
+          >
             <input {...register(id, validation)} id={id} {...getInputProps()} />
             <div
               className={clsx(
                 'w-full p-2 bg-gray-100 border border-gray-300 border-dashed rounded cursor-pointer',
                 errors[id]
-                  ? 'focus:ring-red-500 border-red-500 focus:border-red-500'
-                  : 'focus:ring-dark-400 focus:border-dark-400'
+                  ? 'border-red-500 group-focus:border-red-500'
+                  : 'group-focus:border-primary-500'
               )}
             >
               <p className='my-20 text-center text-gray-500'>
@@ -212,10 +121,15 @@ export default function DropzoneInput({
               <p className='text-sm text-red-500'>{errors[id].message}</p>
             )}
           </div>
-          {!!files?.length && (
-            <ul className='border border-gray-200 divide-y divide-gray-200 rounded-md'>
+          {!readOnly && !!files?.length && (
+            <ul className='border border-gray-300 divide-y divide-gray-300 rounded-md'>
               {files.map((file, index) => (
-                <FilePreview key={index} file={file} deleteFile={deleteFile} />
+                <FilePreview
+                  key={index}
+                  readOnly={readOnly}
+                  file={file}
+                  deleteFile={deleteFile}
+                />
               ))}
             </ul>
           )}
